reorganize and string

+++ libc/string/strlen.c
@@ -1,5 +1,7 @@
-int strlen(const char* s){
-	int i=0;
+#include <string.h>
+
+size_t strlen(const char* s){
+	size_t i=0;
 	while(s[i]!=0){
 		++i;
 	}
